chore: moved chat settings to correct directory
