+++ descendence/src/app/app.module.ts
@@ -23,28 +23,22 @@ import { QrCodeDirective } from './focus-overlay/qrcode.directive';
 import { OnlyNumber } from './focus-overlay/only-numbers.directive';
 import { SharedValidatorService } from './focus-overlay/shared-validator.service';
 import { TwoFactorComponent } from './two-factor/two-factor.component';
-// import { ChatComponent } from './components/chat/chat.component';
 import { ChatComponent } from './chat/chat-client/chat.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { MatInputModule } from '@angular/material/input';
 import { MatListModule } from '@angular/material/list';
 import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
-// import { SearchComponent } from './components/search/search.component';
 import { SearchComponent } from './chat/search/search.component';
 import { SocketIoConfig, SocketIoModule } from 'ngx-socket-io';
-// import { chatGuardService } from './components/chat/chatGuard.service';
 import { chatGuardService } from './chat/chat-client/chatGuard.service';
-// import { CreateChatComponent } from './create-chat/create-chat.component';
 import { CreateChatComponent } from './chat/create-chat/create-chat.component';
 import { MatRadioModule } from '@angular/material/radio';
-// import { ChatPassComponent } from './chat-pass/chat-pass.component';
 import { ChatPassComponent } from './chat/chat-pass/chat-pass.component';
 import { CommonModule } from '@angular/common';
 import { MATERIAL_SANITY_CHECKS } from '@angular/material/core';
-// import { ChatContainerComponent } from './chat-container/chat-container.component';
 import { ChatContainerComponent } from './chat/chat-container/chat-container.component';
-import { SettingsComponent } from './chat/settings.component';
+import { SettingsComponent } from './chat/settings/settings.component';
 
 const config: SocketIoConfig = { url: 'http://localhost:3000', options: { transports: ['websocket']} };
